test(client): fail-closed directory-guard module-edge extraction
Share the comprehensive module-reference extractor, reject non-literal
loads / ImportEquals / require / createRequire escapes, and classify
provider→Runtime seams from normalized resolved targets so
provider-support/../traversal cannot bypass the contracts index allowlist.
