+++ packages/client/src/__tests__/directory-boundary-guard.test.ts
@@ -9,17 +9,17 @@ import {
   writeFileSync,
 } from "node:fs";
 import { tmpdir } from "node:os";
-import { dirname, join, relative } from "node:path";
+import { dirname, join, normalize, relative } from "node:path";
 import { fileURLToPath } from "node:url";
-import ts from "typescript";
 import { afterEach, describe, expect, it } from "vitest";
+import { extractModuleReferences } from "./module-reference-extract.js";
 
 const here = dirname(fileURLToPath(import.meta.url));
 const clientSrc = join(here, "..");
 
 type Layer = "cloud" | "runtime" | "providers" | "root";
 type ResolvedTarget =
-  | { kind: "layer"; layer: Layer }
+  | { kind: "layer"; layer: Layer; resolvedRel: string | null }
   | { kind: "external" }
   | { kind: "unresolved" }
   | { kind: "asset" };
@@ -31,6 +31,14 @@ type ResolvedTarget =
  * modules.
  */
 const RELATIVE_NON_CODE_ASSET_ALLOWLIST: ReadonlySet<string> = new Set([]);
+
+/** Authoritative provider→Runtime seams (matches provider-boundary-guard). */
+const ALLOWED_PROVIDER_RUNTIME_SEAMS = new Set(["runtime/contracts.ts", "runtime/provider-support/index.ts"]);
+
+const ALLOWED_PROVIDER_RUNTIME_PACKAGE_SPECS = new Set([
+  "@first-tree/client/runtime/contracts",
+  "@first-tree/client/runtime/provider-support",
+]);
 
 function walkProductionTs(dir: string, out: string[] = []): string[] {
   for (const name of readdirSync(dir)) {
@@ -51,30 +59,10 @@ function layerOf(srcRoot: string, file: string): Layer | null {
   return null;
 }
 
-function extractSpecifiers(source: string): string[] {
-  const sf = ts.createSourceFile("x.ts", source, ts.ScriptTarget.Latest, true, ts.ScriptKind.TS);
-  const out: string[] = [];
-  const visit = (node: ts.Node) => {
-    if (ts.isImportDeclaration(node) && node.moduleSpecifier && ts.isStringLiteralLike(node.moduleSpecifier)) {
-      out.push(node.moduleSpecifier.text);
-    } else if (ts.isExportDeclaration(node) && node.moduleSpecifier && ts.isStringLiteralLike(node.moduleSpecifier)) {
-      out.push(node.moduleSpecifier.text);
-    } else if (ts.isCallExpression(node) && node.expression.kind === ts.SyntaxKind.ImportKeyword) {
-      const arg = node.arguments[0];
-      if (arg && ts.isStringLiteralLike(arg)) out.push(arg.text);
-    } else if (ts.isImportTypeNode(node) && node.argument && ts.isLiteralTypeNode(node.argument)) {
-      const lit = node.argument.literal;
-      if (ts.isStringLiteralLike(lit)) out.push(lit.text);
-    }
-    ts.forEachChild(node, visit);
-  };
-  visit(sf);
-  return out;
-}
-
+/** Resolve a relative specifier to an existing `.ts` file using normalized path traversal. */
 function resolveRelativeTarget(fromFile: string, spec: string): string | null {
   const stripped = spec.replace(/\.js$/, "");
-  const base = join(dirname(fromFile), stripped);
+  const base = normalize(join(dirname(fromFile), stripped));
   const candidates = [`${base}.ts`, join(base, "index.ts")];
   for (const c of candidates) {
     if (existsSync(c)) return c;
@@ -82,13 +70,32 @@ function resolveRelativeTarget(fromFile: string, spec: string): string | null {
   return null;
 }
 
-function resolveImport(srcRoot: string, fromFile: string, spec: string): ResolvedTarget {
-  if (spec === "@first-tree/client") return { kind: "layer", layer: "root" };
-  if (spec.startsWith("@first-tree/client/cloud")) return { kind: "layer", layer: "cloud" };
-  if (spec.startsWith("@first-tree/client/runtime")) return { kind: "layer", layer: "runtime" };
-  if (spec.startsWith("@first-tree/client/providers")) return { kind: "layer", layer: "providers" };
+function packageSpecLayer(spec: string): Layer | null {
+  if (spec === "@first-tree/client") return "root";
+  if (spec.startsWith("@first-tree/client/cloud")) return "cloud";
+  if (spec.startsWith("@first-tree/client/runtime")) return "runtime";
+  if (spec.startsWith("@first-tree/client/providers")) return "providers";
   // Compatibility export — same ownership as cloud/observability.
-  if (spec.startsWith("@first-tree/client/observability")) return { kind: "layer", layer: "cloud" };
+  if (spec.startsWith("@first-tree/client/observability")) return "cloud";
+  return null;
+}
+
+function resolveImport(srcRoot: string, fromFile: string, spec: string): ResolvedTarget {
+  const pkgLayer = packageSpecLayer(spec);
+  if (pkgLayer) {
+    let resolvedRel: string | null = null;
+    if (spec === "@first-tree/client/runtime/contracts") resolvedRel = "runtime/contracts.ts";
+    else if (
+      spec === "@first-tree/client/runtime/provider-support" ||
+      spec === "@first-tree/client/runtime/provider-support/index"
+    ) {
+      resolvedRel = "runtime/provider-support/index.ts";
+    } else if (spec === "@first-tree/client") resolvedRel = "index.ts";
+    else if (spec.startsWith("@first-tree/client/observability")) {
+      resolvedRel = "cloud/observability/index.ts";
+    }
+    return { kind: "layer", layer: pkgLayer, resolvedRel };
+  }
   if (!spec.startsWith(".")) return { kind: "external" };
 
   const fromRel = relative(srcRoot, fromFile).replaceAll("\\", "/");
@@ -99,17 +106,17 @@ function resolveImport(srcRoot: string, fromFile: string, spec: string): Resolve
   if (!resolved) return { kind: "unresolved" };
   const layer = layerOf(srcRoot, resolved);
   if (!layer) return { kind: "unresolved" };
-  return { kind: "layer", layer };
+  return {
+    kind: "layer",
+    layer,
+    resolvedRel: relative(srcRoot, resolved).replaceAll("\\", "/"),
+  };
 }
 
-function isContractsOrProviderSupport(spec: string): boolean {
-  const normalized = spec.replaceAll("\\", "/");
-  return (
-    /\/runtime\/contracts(?:\.js)?$/.test(normalized) ||
-    normalized.includes("/runtime/provider-support/") ||
-    normalized.endsWith("/runtime/provider-support") ||
-    normalized.endsWith("/runtime/provider-support.js")
-  );
+function isAllowedProviderRuntimeSeam(spec: string, resolvedRel: string | null): boolean {
+  if (ALLOWED_PROVIDER_RUNTIME_PACKAGE_SPECS.has(spec)) return true;
+  if (resolvedRel && ALLOWED_PROVIDER_RUNTIME_SEAMS.has(resolvedRel)) return true;
+  return false;
 }
 
 function collectDirectoryBoundaryViolations(srcRoot: string): string[] {
@@ -123,7 +130,12 @@ function collectDirectoryBoundaryViolations(srcRoot: string): string[] {
     }
 
     const source = readFileSync(file, "utf8");
-    for (const spec of extractSpecifiers(source)) {
+    const refs = extractModuleReferences(source);
+    if (refs.hasUnresolvableModuleReference) {
+      violations.push(`${rel}: unclassifiable/non-literal module reference`);
+    }
+
+    for (const spec of refs.literalSpecifiers) {
       const to = resolveImport(srcRoot, file, spec);
 
       if (to.kind === "external" || to.kind === "asset") continue;
@@ -150,8 +162,12 @@ function collectDirectoryBoundaryViolations(srcRoot: string): string[] {
       if (from === "providers" && to.layer === "cloud") {
         violations.push(`${rel} -> ${spec} (${to.layer})`);
       }
-      if (from === "providers" && to.layer === "runtime" && !isContractsOrProviderSupport(spec)) {
-        violations.push(`${rel} -> ${spec} (runtime outside contracts/provider-support)`);
+      if (from === "providers" && to.layer === "runtime") {
+        if (!isAllowedProviderRuntimeSeam(spec, to.resolvedRel)) {
+          violations.push(
+            `${rel} -> ${spec} (runtime outside contracts/provider-support index; resolved=${to.resolvedRel ?? "n/a"})`,
+          );
+        }
       }
     }
   }
@@ -172,11 +188,14 @@ function makeFixtureSrc(): string {
   fixtureRoots.push(root);
   const src = join(root, "src");
   mkdirSync(join(src, "cloud"), { recursive: true });
-  mkdirSync(join(src, "runtime"), { recursive: true });
+  mkdirSync(join(src, "runtime", "provider-support"), { recursive: true });
   mkdirSync(join(src, "providers"), { recursive: true });
   writeFileSync(join(src, "index.ts"), `export * from "./cloud/index.js";\n`);
   writeFileSync(join(src, "cloud", "index.ts"), `export const cloud = true;\n`);
   writeFileSync(join(src, "runtime", "index.ts"), `export const runtime = true;\n`);
+  writeFileSync(join(src, "runtime", "contracts.ts"), `export type Contract = string;\n`);
+  writeFileSync(join(src, "runtime", "session-manager.ts"), `export class SessionManager {}\n`);
+  writeFileSync(join(src, "runtime", "provider-support", "index.ts"), `export const support = true;\n`);
   writeFileSync(join(src, "providers", "index.ts"), `export const providers = true;\n`);
   return src;
 }
@@ -264,6 +283,58 @@ describe("client directory production dependency direction", () => {
       violations.some((v) => v.includes("providers/missing.ts") && v.includes("unresolved relative TypeScript module")),
     ).toBe(true);
     expect(violations.some((v) => v.includes("cloud/external.ts"))).toBe(false);
+  });
+
+  it("negative fixture: non-literal dynamic import fails closed", () => {
+    const src = makeFixtureSrc();
+    writeFileSync(
+      join(src, "providers", "dyn.ts"),
+      `const target = "../runtime/session-manager.js";\nexport async function load() {\n  return import(target);\n}\n`,
+    );
+    const violations = collectDirectoryBoundaryViolations(src);
+    expect(violations.some((v) => v.includes("providers/dyn.ts") && v.includes("unclassifiable/non-literal"))).toBe(
+      true,
+    );
+  });
+
+  it("negative fixture: ImportEquals require and createRequire binder loads fail closed when targeting Runtime", () => {
+    const src = makeFixtureSrc();
+    writeFileSync(
+      join(src, "providers", "equals.ts"),
+      `import Session = require("../runtime/session-manager.js");\nexport { Session };\n`,
+    );
+    writeFileSync(
+      join(src, "providers", "cjs.ts"),
+      `import { createRequire } from "node:module";\nconst req = createRequire(import.meta.url);\nexport const Session = req("../runtime/session-manager.js");\n`,
+    );
+    const violations = collectDirectoryBoundaryViolations(src);
+    expect(
+      violations.some(
+        (v) => v.includes("providers/equals.ts") && v.includes("runtime outside contracts/provider-support"),
+      ),
+    ).toBe(true);
+    expect(
+      violations.some(
+        (v) => v.includes("providers/cjs.ts") && v.includes("runtime outside contracts/provider-support"),
+      ),
+    ).toBe(true);
+  });
+
+  it("negative fixture: provider-support/../session-manager traversal fails from resolved target", () => {
+    const src = makeFixtureSrc();
+    writeFileSync(
+      join(src, "providers", "traverse.ts"),
+      `import { SessionManager } from "../runtime/provider-support/../session-manager.js";\nexport { SessionManager };\n`,
+    );
+    const violations = collectDirectoryBoundaryViolations(src);
+    expect(
+      violations.some(
+        (v) =>
+          v.includes("providers/traverse.ts") &&
+          v.includes("runtime outside contracts/provider-support") &&
+          v.includes("resolved=runtime/session-manager.ts"),
+      ),
+    ).toBe(true);
   });
 
   it("negative fixture: does not mutate the real packages/client production tree", () => {
